autogen.sh: make building from release tags work
This adapts Brent's commit
https://github.com/libressl/portable/commit/c95574be9296988962cb1358c969144dc382061b
to rpki-client. OPENBSD_BRANCH can contain a branch, a tag, or a commit
hash from the rpki-client-openbsd repo, and git will do the right thing.

To make this actually work, we will need to create a release commit
where OPENBSD_BRANCH contains 'rpki-client-9.7'. This commit could
happen on master if we want to insist on a linear history or it could
live on a dedicated branch. I don't have a particular preference here
except that extra branches could be backported to earlier releases if
we care to do so.

Fixes #120 (starting with release 9.7)